services/center_client/requirements.txt
websockets==12.0
pyyaml==6.0.1
asyncio-mqtt==0.16.1
